remove monolithic auth_flows module
