+++ pkgs/standards/auto_authn/auto_authn/v2/routers/auth_flows/common.py
@@ -1,0 +1,42 @@
+from __future__ import annotations
+
+from typing import Optional
+
+from fastapi import APIRouter, HTTPException, Request, status
+from pydantic import BaseModel, Field
+
+from ...jwtoken import JWTCoder
+from ...backends import ApiKeyBackend, PasswordBackend
+from ...runtime_cfg import settings
+
+
+router = APIRouter()
+
+_jwt = JWTCoder.default()
+_pwd_backend = PasswordBackend()
+_api_backend = ApiKeyBackend()
+
+AUTH_CODES: dict[str, dict] = {}
+SESSIONS: dict[str, dict] = {}
+
+
+def _require_tls(request: Request) -> None:
+    if settings.require_tls and request.url.scheme != "https":
+        raise HTTPException(status.HTTP_400_BAD_REQUEST, {"error": "tls_required"})
+
+
+async def _front_channel_logout(session_id: str) -> None:
+    """Placeholder for front-channel logout notifications."""
+    return None
+
+
+async def _back_channel_logout(session_id: str) -> None:
+    """Placeholder for back-channel logout notifications."""
+    return None
+
+
+class TokenPair(BaseModel):
+    access_token: str
+    refresh_token: str
+    token_type: str = Field(default="bearer")
+    id_token: Optional[str] = None
